Move process_state.hpp
