+++ score/launch_manager/src/daemon/src/configuration/configuration_adapter.hpp
@@ -17,7 +17,7 @@
 #include "score/mw/launch_manager/common/constants.hpp"
 #include "score/mw/launch_manager/common/identifier_hash.hpp"
 #include "score/mw/launch_manager/common/process_group_state_id.hpp"
-#include "score/mw/launch_manager/common/process_state.hpp"
+#include "score/mw/launch_manager/process_group_manager/process_state.hpp"
 #include "score/mw/launch_manager/configuration/config.hpp"
 #include "score/mw/launch_manager/process_group_manager/iprocess.hpp"
 #include <chrono>
